moving ai llm interaction to ai/ai_client
